Do some code splitting to make future work on flakehub-push more joyful
